Use radians for angle + use cannon position for lasers
By using radians for the angle, we no longer have to do a conversion and be
mindful of two different systems.

The cannon is a fixed point on a sprite, so we take that node's
`global_position` to get where the cannon should be after all the rotations and
movement. Luckily this trick works fine.

A couple of issues remain for the lasers:

* Bug: lasers don't shoot forward when player is moving
* Feature: only one laser should be shot per second