Player reacts to virus and laser collisions
If the virus directly collides with a player or a laser hits the player,
the player's takes their damage and reduces their health.

We can't do typecasting or the like as I do with Phaser.
Since it's Python-like, we can just check if a property exists.
This isn't a bad way to get damage from a generic Area2D node.

We can always add identifier properties for games with larger scope.

When a player's health is 0, it should emit the died signal for game over.

A player can listen to it's own signal, which is amazing.
Anything that affects our health can force us to check if we're alive.
This is a VERY scalable solution. Kudos Godot!