Create an identifier scene
While managing collisions, adding a node_type isn't enough.
The player collisions needs to not collide with its own lasers.
It also needs to collide with other viruses and their lasers.

So two different scenes need to be identified: lasers + viruses.
It makes more sense to create this node with an ID.
Those 2 scenes can then make it a child, instead of maintaining a
node_type property.