Let's add 10 the score after killing an enemy
This is just to confirm the connection between HUD and main works.

As we instance viruses via the script, we need to use the `connect`
function to get signals.
The signal name does not begin with an _ to denote its developer made.

Would love to add points based on the virus property itself