Add a start timer to begin game
We allow the begin time to be configurable, 3 seconds is fine.

To get a countdown we have a looping start timer for 1 second.
This detracts from our main node's wait time.
When it reaches 0, we kick off the game.

We also position the player lower in the screen to be visible